pyproject.toml
[project]
name = "repl-nix-cvanalyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "nltk>=3.9.1",
    "googletrans==3.1.0a0",
    "openai>=0.28.1",
    "reportlab",
]
